+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Marshaller that uses the name of the struct (without the package)
+
 
 ## [0.1.0] - 2019-03-16